+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### [0.0.2]
-- style config expansion (added copyWith for config classes and customization for Calendar()
+* style config expansion (added `copyWith` for config classes, customization for Calendar()
 and RangeCalendar() background color, radius, header text style, border radius of date range
 in RangeCalendar())
 ### [0.0.1]
-- init
+* init